requirements.txt
asgiref==3.8.1
channels==4.1.0
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
PyMySQL==1.1.1
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2
whitenoise==6.6.0
